requirements.txt
scanpy>=1.9.6
torch>=2.0.0
anndata>=0.10.3
numpy>=1.23.5
setuptools>=59.5.0
pandas>=2.1.1
matplotlib>=3.8.1
scikit-learn>=1.3.2
tqdm>=4.66.1
biolord>=0.1.0
seaborn>=0.12.2
scipy>=1.11.3
scvi-tools>=0.16.1
pytorch-lightning>=1.5.10
